fix bugs

+++ habitat_extensions/sensors.py
@@ -90,7 +90,7 @@ class ShortestPathSensor(Sensor):
         return SensorTypes.TACTILE
 
     def _get_observation_space(self, *args: Any, **kwargs: Any):
-        return spaces.Box(low=0.0, high=100, shape=(1,), dtype=np.float)
+        return spaces.Box(low=0.0, high=100, shape=(1,), dtype=np.float32)
 
     def get_observation(self, *args: Any, episode, **kwargs: Any):
         best_action = self.follower.get_next_action(episode.goals[0].position)
@@ -128,7 +128,7 @@ class VLNOracleProgressSensor(Sensor):
         return SensorTypes.MEASUREMENT
 
     def _get_observation_space(self, *args: Any, **kwargs: Any):
-        return spaces.Box(low=0.0, high=1.0, shape=(1,), dtype=np.float)
+        return spaces.Box(low=0.0, high=1.0, shape=(1,), dtype=np.float32)
 
     def get_observation(
         self, observations, *args: Any, episode, **kwargs: Any
